set up the thing to work on my pc
ok so first there was a problem that was it tried running `runserver_plus`
instead of `runserver`, so I fixed that.
Then it disliked the quotes in the .django files, so I removed those.
Finally, since my Docker requires me to run it in a VM, I had to include
an extra IP as valid. Now it works, so I can work on the website finally