Fully automated setup of all Window Manager functions.
When created a window, it is no longer neccesary to manually assign an 
Id to each window, or to assign onclick functions to any of the Window 
Buttons. 

This makes it much easier to create new windows, both manually and with 
scripts, as there is much less boilerplate code, and the windows' ids 
are ensured to be unique.