Clarify template logic and fix url inconsistency
The new user_detail.html and webpage_list.html account for the default
value of a Webpage object - with 'placeholder_img' in the path -
not just a blanket check for the url (which will always be True).

For Template-object-related url templates,
one url used the pk while another used the templatename.
Both now use templatename.
(since this field is now unique in the Template model)