Support responses for whether the request is POST or GET
Originally, the register page would always return itself upon
submission. This was problematic as:

If the form submission was a POST request then:
	- The form was not saved.
	- The form was not checked for validity.
	- The registration form would be confusingly returned again.
Now, if the method is POST, the form is checked to be valid, and if it
is, then it's saved, and the user's redirected back to home.
Any other cases have the registration form returned again.