Improve basic page functionality
Add '?next={{ request.path }}' to end of urls to ensure user is
redirected back to the page where they clicked the button. More of a
convenience feature than anything!

Added clarification to user_detail.html so that the page says if the
user hasn't created any pages.

Corrected typo in register.html.