Provide an initial config and let it interact with docker
I ran into an issue with `allowed_hosts`, and I had to figure out how
to create a `config.json` (as in, what format), and had to edit the
dockerignore... Not very friendly.

So, this commit just allows null keys, which means that the key will get generated for you!