Renamed 'city' to 'place', and 'profile_css' to 'custom_css'.
We decided that 'place' is a broader term and it means that people can be creative with their profiles without the need to validate if a city is actually real.
The blog and profile will share the same custom CSS, so it makes more sense for it to be named 'custom_css' instead.