Remove the use of the Mars weather API
At the time of writing this, the weather API is only returning 5 days of 
weather when our weekly weather data anticipates 7. This is causing an 
issue making the game unplayable. These changes removes the use of the 
weekly weather data to use a default values for the `wind_speed` and 
`temperature` variables.